Mackey contexts

One of our time-saving design choices is the idea of a MackeyContext, which precomputes relevant information about a finite group before working with Mackey functors for that group. The idea is to make the construction of a Mackey functor a two-step process: step 1 is to build the context, and then step 2 is to build a Mackey functor for the context. This allows us to avoid certain costly computations, for instance computing the double coset formulas.

Definition Let $H\le K\le G$ be subgroups of a group $G$. We say that the tuple $(H,K)$ is a cover (and we will denote this by $H\lesssim K$) if $H$ is a proper subgroup of $K$, and if $H\le L \le K$ is a pair of subgroup inclusions, then either $H=L$ or $L=K$.

The MackeyContext data type stores the following information about the group:

the group itself

a list of all subgroups of the group

a list of all the covers $H\lesssim K$

a list of paths between two subgroups: for every tuple of subgroups $(H,K)$ with $H\le K$, we store a single sequence of covers $H = H_0 \lesssim H_1 \lesssim \cdots \lesssim H_n = K$

a list of generators for the group

a list of relation words among those generators

a matrix generator_left_conjugation_matrix, whose rows are indexed by generators of the group and whose columns are indexed by subgroups, and where the $[g,H]$th entry is the index in the list of subgroups of the subgroup $gHg^{-1}$

a similar matrix generator_right_conjugation_matrix telling us how to find $g^{-1}Hg$ for a generator $g$

a dictionary double_coset_formulae: the keys are triples of subgroups $(J,H,K)$, with $J\lesssim H$ and $K\lesssim H$, and the values are a vector of DoubleCosetFormulaTerm types: these are tuples of generator words and subgroup indices, intended to help us locate $J^x \cap K$.

Proposition: Let $M$ give the data of $M(H)$ for every subgroup $H\le G$, a conjugation map $c_g \colon M(H) \to M(gHg^{-1})$ for every generator $g\in G$, and restriction and transfer maps for every cover $H\lesssim K$. Then if the double coset formula holds along covers, it holds in general.

MackeyFunctors.MackeyContext — Type

MackeyContext(G)

Precompute subgroup-lattice and double-coset data for a finite GAP group G. This front-loads the computational effort for the group and subgroup-lattice data used when verifying the axioms for any $G$-Mackey functor.

This method takes a group $G$ and precomputes its list of subgroups, the list of covers (meaning proper subgroups $H\le K$ where there are no intermediate subgroups), a list of generators for the group, the relators for those generators, and matrices for how these generators act via conjugation on the list of all subgroups.

Importantly, the MackeyContext type also stores all the data needed to verify the double coset formula for an input Mackey functor. An important lemma is that the double coset formula can be checked along subgroups $J\le H \ge K$ where each inclusion is a cover. Given a triple of subgroups as above, we have

\[H = \coprod_x JxK\]

and we store this in a dictionary as a vector of entries $(J^x\cap K, x)$.