Constructors

MackeyFunctors.constant_mackey_functor — Function

constant_mackey_functor(ctx,M)

Given a MackeyContext and an $R$-module $M$, this method outputs the fixed-point Mackey functor for $M$ with trivial $G$-action. This is also called the constant Mackey functor valued at $M$.

constant_mackey_functor(ctx,R)

This method can also be fed a context and a ring $R$, and it will output the constant Mackey functor valued at $R$ considered as a free rank one $R$-module.

MackeyFunctors.burnside_mackey_functor — Function

burnside_mackey_functor(mc::MackeyContext, R::Ring) -> MackeyFunctor

Return the Burnside Mackey functor for the group specified by mc and the coefficient ring R.

burnside_mackey_functor(G, R::Ring) -> MackeyFunctor

Return the Burnside Mackey functor for the group G and the coefficient ring R.