Mackey functors

When asking a user to construct a Mackey functor for a group $G$, it seems reasonable to ask them to supply the values $M(H)$ for each subgroup $H\le G$, but asking for restriction/transfer along every subgroup inclusion and conjugation for every group element feels a bit excessive. A more concise collection of data is:

the values $M(H)$ for each $H\le G$

restriction and transfer along covers

conjugation by generators of $G$

This is a compression of the data of a Mackey functor available in Webb. Obviously steps are needed to verify that this does indeed satisfy all the axioms of a Mackey functor. Some checks that need to be verified include:

Check 1: If $h\in H$, then the conjugation map $c_h$ is the identity on $M(H)$.

Check 2: The conjugation maps for generators of $G$ respect the relations in $G$.

Check 3: Conjugation by generators commutes with restrictions and transfers along covers.

Check 4: For any two factorizations of $H \le K$ into covers, the supplied cover restriction maps and cover transfer maps yield a well-defined restriction/transfer between $M(H)$ and $M(K)$.

Check 5: The double coset formula holds for triples $(J,H,K)$ where $J\lesssim H$ and $K\lesssim H$ are covers.

Theorem: The data supplied above, satisfying all five checks, uniquely specifies a $G$-Mackey functor.

MackeyFunctors.MackeyFunctor — Type

MackeyFunctor(G)

todo

Extracting data from a Mackey functor

Given a Mackey functor, we can access various parts of the data.

AbstractAlgebra.coefficient_ring — Function

coefficient_ring(M)

Returns the underlying coefficient ring of the Mackey functor M.

MackeyFunctors.transfer — Function

transfer(M,i,j)

Given a Mackey functor M and two SubgroupIndex values i and j corresponding to subgroups $H = H[i] \le H[j] = K \le G$, this method returns the transfer homomorphism $M(H) \to M(K)$ as the type AbstractAlgebra.Generic.ModuleHomomorphism.

MackeyFunctors.restriction — Function

restriction(M,i,j)

Given a Mackey functor M and two SubgroupIndex values i and j corresponding to subgroups $H = H[i] \le H[j] = K \le G$, this method returns the restriction homomorphism $M(K) \to M(H)$ as the type AbstractAlgebra.Generic.ModuleHomomorphism.

MackeyFunctors.conjugation — Function

conjugation(M,g,i)

Given a Mackey functor M, a GroupElement $g\in G$, and a SubgroupIndex i, this method returns the conjugation isomorphism $M(H[i])\to M(gH[i]g^{-1})$ as an AbstractAlgebra.Generic.ModuleIsomorphism.

MackeyFunctors.value — Function

value(M,i)

Returns the value of the Mackey functor M at the subgroup index i.